// Neptune.Web/src/app/shared/generated/model/treatment-bmp-basic-info-update.ts
export class TreatmentBMPBasicInfoUpdate { 
    TreatmentBMPName: string;
    OwnerOrganizationID?: number | null;
    YearBuilt?: number | null;
    SystemOfRecordID?: string | null;
    WaterQualityManagementPlanID: number;
    TreatmentBMPLifespanTypeID: number;
    TreatmentBMPLifespanEndDate?: string | null;
    SizingBasisTypeID: number;
    TrashCaptureStatusTypeID: number;
    TrashCaptureEffectiveness?: number | null;
    RequiredFieldVisitsPerYear?: number | null;
    RequiredPostStormFieldVisitsPerYear?: number | null;
    Notes?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface TreatmentBMPBasicInfoUpdateForm { 
    TreatmentBMPName: FormControl<string>;
    OwnerOrganizationID?: FormControl<number>;
    YearBuilt?: FormControl<number>;
    SystemOfRecordID?: FormControl<string>;
    WaterQualityManagementPlanID: FormControl<number>;
    TreatmentBMPLifespanTypeID: FormControl<number>;
    TreatmentBMPLifespanEndDate?: FormControl<string>;
    SizingBasisTypeID: FormControl<number>;
    TrashCaptureStatusTypeID: FormControl<number>;
    TrashCaptureEffectiveness?: FormControl<number>;
    RequiredFieldVisitsPerYear?: FormControl<number>;
    RequiredPostStormFieldVisitsPerYear?: FormControl<number>;
    Notes?: FormControl<string>;
}

export class TreatmentBMPBasicInfoUpdateFormControls { 
    public static TreatmentBMPName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
                Validators.minLength(1),
            ],
        }
    );
    public static OwnerOrganizationID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static YearBuilt = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static SystemOfRecordID = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static WaterQualityManagementPlanID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static TreatmentBMPLifespanTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static TreatmentBMPLifespanEndDate = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static SizingBasisTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static TrashCaptureStatusTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static TrashCaptureEffectiveness = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
                Validators.min(1),
                Validators.max(99),
            ],
        }
    );
    public static RequiredFieldVisitsPerYear = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
                Validators.min(0),
                Validators.max(2147483647),
            ],
        }
    );
    public static RequiredPostStormFieldVisitsPerYear = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
                Validators.min(0),
                Validators.max(2147483647),
            ],
        }
    );
    public static Notes = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
                Validators.minLength(0),
                Validators.maxLength(2000),
            ],
        }
    );
}